Set default cluster load rates and counters to zero
To simplify load statistics aggregation default cluster
load rates and counters should be set to zero when mastermind worker
is starting. This prevents a lot of additional checks
when statistics is being processed.
For example, disk's read rate can be calculated when we process
cluster statistics for the second time. Instead of setting this
value to None we are satistied with the default 0.0Mb/s rate.