Accept generators to GroupNodeBackendsSet
GroupNodeBackendsSet should contain only unique GroupNodeBackendRecord-s,
so a set is being constructed to check this fact during object
construction. If generator is passed it will be exhausted on the first
pass, so the set will get empty sequence of elements.
We prevent this by passing `self` to set constuctor because `self`
is already an iterable.