Account couples on defragmenting disks
When a couple has one group on a disk that has a defragmentation running on
(or one that is about to run in case of a new job), it should
be placed in a separate weight bucket because otherwise it will potentially
respond with a fail to WRITE operation. Such couples should be
taken into account only when weight calculator runs out of 'base' bucket
couples.