Use blob_size_limit as total space unconditionally
Attribute ```total_space``` used to be limited by ```vfs_total_space```
when we had no filesystem sanity check (sum of ```total_space```
of all backends residing on a filesystem should be less than
or equal to ```vfs_total_space``` of that filesystem).
Now this limitation is redundant and can actually hide
misconfiguration issues.