Implement cached key upload queue
Cached key state is stored in mongo and looks something like this:

{
    'id': 'ef0a00111',
    'couple': '1:2:3',
    'cached_groups': [666, 999],
    ...
}

Here 'cached_groups' key is a list of cache groups' ids
where the key is being stored.

Except it's not!

Actually, when mastermind decides to place key into a cache group,
it creates a task for a gatlinggun and immediately adds that cache group's
id to a 'cached_groups' list. This can lead to -2 errors on client's side
when he desides to use cache groups to read a key.

Additionally, we do not resrict the number of copies the are being made
in dc simultaneously.

Solution:
We add intermediate state when the task for gatlinggun has been created
but not yet processed. This state is described by 'add_queue' list:

{
    'id': 'ef0a00111',
    'couple': '1:2:3',
    'cached_groups': [666],
    'add_queue': [
        {
            'cache_group': 999,
            'ts': 1455226998,
        }
    ],
    ...
}

This example tells us that at timestamp 1455226998 the task for gatlinggun
on cache group 999 was added. Note that 'cached_groups' list does not include
group 999 any more. That means that client will not try to read a key
from it until it becomes available.

We also restrict the number of copies in a dc to only one at a time.
This allows us to create additional copies using cache groups
that already have a copy of the key as a source.