Add 'update_cache_key_status' handle
This commit continues the idea of the previous one.

When gatlinggun processes the task of uploading the key,
it should be able to notify mastermind of operation's result.

Take a look at the previous example:
    {
        'id': 'ef0a00111',
        'couple': '1:2:3',
        'cached_groups': [666],
        'add_queue': [
            {
                'cache_group': 999,
                'ts': 1455226998,
            }
        ],
        ...
    }

When the key is successfully uploaded to group 999, gatlinggun
makes 'update_cache_key_status' request telling mastermind
about its great success which results in the cached key state
changing to something like this:

    {
        'id': 'ef0a00111',
        'couple': '1:2:3',
        'cached_groups': [666, 999],
        'add_queue': [],
        ...
    }

Now the client is able to use cache group 999 to read the cached key.

TODO: the analogous workflow when removing the key from cache group