Add dnet_client backend minion task
Right now minion relies on mastermind to supply
shell command that should be executed. But there are
cases when mastermind does not have all information
to construct the command. Like when new group is
created on a file system, mastermind wants to run it
but does not know which backend id is assigned to it
(backend id is usually fetched from monitor stat data
of running backend).

But, fortunatelly, mastermind-minion can fetch the rest
of required data itself because it can read elliptics config
locally on the node and fill in the missing parameters.
This is what new minion task type is for. It allows
mastermind to supply command template and values that
it can provide and let mastermind-minion fetch the rest.
Afterward minion runs the command in the usual way.