Set group's active job when job is linked to a couple
Jobs in mastermind are linked to entities they process.
For most job types jobs where linked to a certain group.
This commit adds the ability to link job to a certain couple
instead of just a group. This can be useful when job processes
couple rather than a separate couple's group.