Skip pending jobs' resources
The list of pending and broken recover dc and defragmentation jobs
happens to be non-empty all the time, and investigation takes a lot of time
while new jobs are being on hold.
As for now, we want to defarg and recover as fast as we can, so broken and
pending jobs' resources should not be accounted for when jobs are being
scheduled.