Allow cooperative running of low priority jobs
Previous approach that we used was to skip running low priority jobs
when any of the higher priority job used any of the required resources.
Since we want to be able to run low priority job along with high priority
ones (for service purposes), we cannot use this approach anymore.

Therefore, resources are now used in the following way:
- every job type has a resource limit usage for resource types that it
requires
- when calculating free resource slots for a job type we consider all
jobs of the same and higher priority and sum its' resource usage
- if resource limit usage for the current job type is greater than
the current resource usage by the jobs of the same and higher priority,
we have free resource slots that can be used.